refactor names of files, add dates parsing via moment.js, change tab of the event tab, minor changes
