+++ package.json
@@ -20,6 +20,7 @@
     "axios": "^0.19.2",
     "ionicons": "^4.6.3",
     "lodash": "^4.17.15",
+    "moment": "^2.24.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router": "^5.1.2",
