+++ src/App.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { Redirect, Route } from 'react-router-dom';
 import { IonApp, IonIcon, IonLabel, IonRouterOutlet, IonTabBar, IonTabButton, IonTabs } from '@ionic/react';
 import { IonReactRouter } from '@ionic/react-router';
-import { apps, flash, send } from 'ionicons/icons';
-import Tab1 from './pages/Tab1';
+import { apps, list, send } from 'ionicons/icons';
+import EventsListPage from './pages/eventsListPage/EventsListPage';
 import Tab2 from './pages/Tab2';
 import Tab3 from './pages/Tab3';
 import Details from './pages/Details';
-import EventDetailPage from './pages/eventPage/eventDetailPage/EventDetailPage';
+import EventDetailPage from './pages/eventsListPage/eventDetailPage/EventDetailPage';
 
 /* Core CSS required for Ionic components to work properly */
 import '@ionic/react/css/core.css';
@@ -34,16 +34,16 @@ const App: React.FC = () => (
             <IonTabs>
                 <IonRouterOutlet>
                     <Route path="/eventDetail/:id" component={EventDetailPage} exact={true} />
-                    <Route path="/tab1" component={Tab1} exact={true} />
+                    <Route path="/eventsList" component={EventsListPage} exact={true} />
                     <Route path="/tab2" component={Tab2} exact={true} />
                     <Route path="/tab2/details" component={Details} />
                     <Route path="/tab3" component={Tab3} />
                     <Route path="/" render={() => <Redirect to="/tab1" />} exact={true} />
                 </IonRouterOutlet>
                 <IonTabBar slot="bottom">
-                    <IonTabButton tab="tab1" href="/tab1">
-                        <IonIcon icon={flash} />
-                        <IonLabel>Tab One</IonLabel>
+                    <IonTabButton tab="eventsList" href="/eventsList">
+                        <IonIcon icon={list} />
+                        <IonLabel>Event list</IonLabel>
                     </IonTabButton>
                     <IonTabButton tab="tab2" href="/tab2">
                         <IonIcon icon={apps} />
